move LinkItemDetailViewModel into binders package
