+++ apple/OmnivoreKit/Sources/Binders/Views/LinkItemDetailView.swift
@@ -1,30 +1,114 @@
 import Combine
 import Models
+import Services
 import SwiftUI
 import Utils
+import Views
 
-public enum PDFProvider {
-  public static var pdfViewerProvider: ((URL, FeedItem) -> AnyView)?
+// TODO: remove this view model
+extension LinkItemDetailViewModel {
+  static func make(feedItem: FeedItem, services: Services) -> LinkItemDetailViewModel {
+    let viewModel = LinkItemDetailViewModel(item: feedItem)
+    viewModel.bind(services: services)
+    return viewModel
+  }
+
+  func bind(services: Services) {
+    performActionSubject.sink { [weak self] action in
+      switch action {
+      case .load:
+        self?.loadWebAppWrapper(services: services)
+      case let .updateReadStatus(markAsRead: markAsRead):
+        self?.updateItemReadStatus(markAsRead: markAsRead, dataService: services.dataService)
+      }
+    }
+    .store(in: &subscriptions)
+  }
+
+  private func updateItemReadStatus(markAsRead: Bool, dataService: DataService) {
+    dataService
+      .updateArticleReadingProgressPublisher(
+        itemID: item.id,
+        readingProgress: markAsRead ? 100 : 0,
+        anchorIndex: 0
+      )
+      .sink { completion in
+        guard case let .failure(error) = completion else { return }
+        print(error)
+      } receiveValue: { [weak self] feedItem in
+        self?.item.readingProgress = feedItem.readingProgress
+      }
+      .store(in: &subscriptions)
+  }
+
+  private func loadWebAppWrapper(services: Services) {
+    // Attempt to get `Viewer` from DataService
+    if let currentViewer = services.dataService.currentViewer {
+      createWebAppWrapperViewModel(username: currentViewer.username, services: services)
+      return
+    }
+
+    services.dataService.viewerPublisher().sink(
+      receiveCompletion: { completion in
+        guard case let .failure(error) = completion else { return }
+        print(error)
+      },
+      receiveValue: { [weak self] viewer in
+        self?.createWebAppWrapperViewModel(username: viewer.username, services: services)
+      }
+    )
+    .store(in: &subscriptions)
+  }
+
+  private func createWebAppWrapperViewModel(username: String, services: Services) {
+    let baseURL = services.dataService.appEnvironment.webAppBaseURL
+
+    let urlRequest = URLRequest.webRequest(
+      baseURL: services.dataService.appEnvironment.webAppBaseURL,
+      urlPath: "/app/\(username)/\(item.slug)",
+      queryParams: ["isAppEmbedView": "true", "highlightBarDisabled": isMacApp ? "false" : "true"]
+    )
+
+    let newWebAppWrapperViewModel = WebAppWrapperViewModel(
+      webViewURLRequest: urlRequest,
+      baseURL: baseURL,
+      rawAuthCookie: services.authenticator.omnivoreAuthCookieString
+    )
+
+    newWebAppWrapperViewModel.performActionSubject.sink { action in
+      switch action {
+      case let .shareHighlight(highlightID):
+        print("show share modal for highlight with id: \(highlightID)")
+      }
+    }
+    .store(in: &newWebAppWrapperViewModel.subscriptions)
+
+    webAppWrapperViewModel = newWebAppWrapperViewModel
+  }
 }
 
-public final class LinkItemDetailViewModel: ObservableObject {
-  @Published public var item: FeedItem
-  @Published public var webAppWrapperViewModel: WebAppWrapperViewModel?
+enum PDFProvider {
+  static var pdfViewerProvider: ((URL, FeedItem) -> AnyView)?
+}
 
-  public enum Action {
+final class LinkItemDetailViewModel: ObservableObject {
+  @Published var item: FeedItem
+  @Published var webAppWrapperViewModel: WebAppWrapperViewModel?
+
+  enum Action {
     case load
     case updateReadStatus(markAsRead: Bool)
   }
 
-  public var subscriptions = Set<AnyCancellable>()
-  public let performActionSubject = PassthroughSubject<Action, Never>()
+  var subscriptions = Set<AnyCancellable>()
+  let performActionSubject = PassthroughSubject<Action, Never>()
 
-  public init(item: FeedItem) {
+  init(item: FeedItem) {
     self.item = item
   }
 }
 
-public struct LinkItemDetailView: View {
+struct LinkItemDetailView: View {
   @Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
 
   static let navBarHeight = 50.0
@@ -32,7 +116,7 @@ public struct LinkItemDetailView: View {
   @State private var showFontSizePopover = false
   @State private var navBarVisibilityRatio = 1.0
 
-  public init(viewModel: LinkItemDetailViewModel) {
+  init(viewModel: LinkItemDetailViewModel) {
     self.viewModel = viewModel
   }
 
@@ -61,7 +145,7 @@ public struct LinkItemDetailView: View {
     )
   }
 
-  public var body: some View {
+  var body: some View {
     #if os(iOS)
       if UIDevice.isIPhone, !viewModel.item.isPDF {
         compactInnerBody
